requirements.txt
altgraph==0.17.2
bottle==0.12.19
bottle-websocket==0.2.9
cffi==1.15.0
Eel==0.14.0
et-xmlfile==1.1.0
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.2
openpyxl==3.0.9
pefile==2021.9.3
Pillow==8.4.0
pycparser==2.21
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
pyparsing==3.0.6
pywin32-ctypes==0.2.0
tinyaes==1.0.2
unicode==2.8
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
